DRY up the scholarship info block a bit
